% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahr_blinded.r
\name{ahr_blinded}
\alias{ahr_blinded}
\title{Blinded estimation of average hazard ratio}
\usage{
ahr_blinded(
  Srv = Surv(time = simtrial::Ex1delayedEffect$month, event =
    simtrial::Ex1delayedEffect$evntd),
  intervals = array(3, 3),
  hr = c(1, 0.6),
  ratio = 1
)
}
\arguments{
\item{Srv}{input survival object (see \code{Surv}); note that only 0=censored, 1=event for \code{Surv}}

\item{intervals}{Vector containing positive values indicating interval lengths where the
exponential rates are assumed.
Note that a final infinite interval is added if any events occur after the final interval
specified.}

\item{hr}{vector of hazard ratios assumed for each interval}

\item{ratio}{ratio of experimental to control randomization.}
}
\value{
A \code{tibble} with one row containing
`AHR` blinded average hazard ratio based on assumed period-specific hazard ratios input in `failRates`
and observed events in the corresponding intervals
`Events` total observed number of events, `info` statistical information based on Schoenfeld approximation,
and info0 (information under related null hypothesis) for each value of `totalDuration` input;
if `simple=FALSE`, `Stratum` and `t` (beginning of each constant HR period) are also returned
and `HR` is returned instead of `AHR`
}
\description{
Based on blinded data and assumed hazard ratios in different intervals, compute
a blinded estimate of average hazard ratio (AHR) and corresponding estimate of statistical information.
This function is intended for use in computing futility bounds based on spending assuming
the input hazard ratio (hr) values for intervals specified here.
}
\examples{
library(simtrial)
library(survival)
ahr_blinded(Srv = Surv(time = simtrial::Ex2delayedEffect$month,
                       event = simtrial::Ex2delayedEffect$evntd),
            intervals = c(4,100),
            hr = c(1, .55),
            ratio = 1)

}
