% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt.fixed_design}
\alias{as_gt.fixed_design}
\title{This is the function to format the bounds summary table of fixed design into gt style.}
\usage{
\method{as_gt}{fixed_design}(x, title = NULL, footnote = NULL)
}
\arguments{
\item{x}{}

\item{title}{}

\item{footnote}{}
}
\value{

}
\description{
This is the function to format the bounds summary table of fixed design into gt style.
}
\examples{
fixed_design(x = "AHR", 
             alpha = alpha,  
             enrollRates = enrollRates, 
             failRates = failRates, 
             studyDuration = studyDuration, 
             ratio = ratio
             ) \%>\% 
             summary()
}
