% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt}
\alias{as_gt}
\title{This is the function to format the bounds summary table into gt style.}
\usage{
as_gt(
  x,
  title = NULL,
  subtitle = NULL,
  colname_spanner = "Cumulative boundary crossing probability",
  colname_spannersub = c("Alternate hypothesis", "Null hypothesis"),
  footnote = NULL
)
}
\arguments{
\item{x}{an object returned by \code{summary_bound}}

\item{title}{a string to specify the title of the gt table}

\item{subtitle}{a string to specify the subtitle of the gt table}

\item{colname_spanner}{a string to specify the spanner of the gt table}

\item{colname_spannersub}{a vector of strings to specify the spanner details of the gt table}

\item{footnote}{a list containing \code{content}, \code{location}, and \code{attr}.
the \code{content} is a vector of string to specify the footnote text;
the \code{location} is a vector of string to specify the locations to put the superscript of the footnote index;
the \code{attr} is a vector of string to specify the attributes of the footnotes, e.g., c("colname", "title", "subtitle", "analysis", "spanner");
users can use the functions in the \code{gt} package to custom themselves.}
}
\value{
a gt table summarizing the bounds table in group sequential designs
}
\description{
This is the function to format the bounds summary table into gt style.
}
\examples{
x_combo <- gsdmvn::gs_design_combo2(
  ratio = 1, 
  alpha = 0.025, 
  beta = 0.2, 
  enrollRates = tibble::tibble(Stratum = "All", duration = 12, rate = 500/12),
  failRates = tibble::tibble(Stratum = "All", duration = c(4, 100), 
                             failRate = log(2) / 15, hr = c(1, .6), dropoutRate = .001), 
  fh_test = fh_test,
  upper = gsdmvn::gs_spending_combo,
  upar = list(sf = gsDesign::sfLDOF, total_spend = 0.025),
  lower = gsdmvn::gs_spending_combo,
  lpar = list(sf = gsDesign::sfLDOF, total_spend = 0.2))
  summary_bound(x_combo2, method = "COMBO") \%>\% 
as_gt(
  title = "Summary of the Crossing Probability",
  subtitle = "by Using gs_design_wlr",
  colname_spanner = "Cumulative boundary crossing probability",
  colname_spannersub = c("Alternate hypothesis", "Null hypothesis"),
  footnote = list(content = c("gs_design_wlr is a function in gsdmvn.",
                              "EF is the events fraction, i.e., Events/N."),
                  location = c(NA, NA),
                  attr = c("subtitle", "analysis")))
                          
}
