% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_design.R
\name{fixed_design}
\alias{fixed_design}
\title{Fixed design sample size}
\usage{
fixed_design(
  x = "AHR",
  alpha = 0.025,
  power = 0.9,
  ratio = 1,
  studyDuration = 36,
  ...
)
}
\arguments{
\item{x}{Sample size method; default is "AHR"; see examples and details.}

\item{alpha}{One-sided Type I error (strictly between 0 and 1)}

\item{power}{Power (\code{NULL} to compute power or strictly between 0 and \code{1 - alpha} otherwise)}

\item{ratio}{Experimental:Control randomization ratio}

\item{studyDuration}{study duration}

\item{...}{}
}
\value{

}
\description{
Computes fixed design sample size for many sample size methods.
Returns a \code{tibble} with a basic summary
}
\examples{
# Average hazard ratio
y <- fixed_design("AHR", 
                  alpha = .025, Power = .9, 
                  enrollRates = tibble::tibble(Stratum = "All",  duration = 18, rate = 1),
                  failRates = tibble::tibble(Stratum = "All", duration = c(4, 100), failRate = log(2) / 12, hr = c(1, .6), dropoutRate = .001),
                  studyDuration = 36)
y \%>\% summary_fix()            

# Lachin and Foulkes (uses gsDesign::nSurv())
y <- fixed_design("LF", 
                  alpha = .025, Power = .9, 
                  enrollRates = tibble::tibble(Stratum = "All",  duration = 18, rate = 1),
                  failRates = tibble::tibble(Stratum = "All", duration = 100, failRate = log(2) / 12, hr = .7, dropoutRate = .001),
                  studyDuration = 36)

}
