% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_design_rmst.R
\name{fixed_design_power_rmst}
\alias{fixed_design_power_rmst}
\title{Power calculation based on RMST method}
\usage{
fixed_design_power_rmst(
  enrollRates,
  failRates,
  analysisTimes,
  ratio = 1,
  alpha = 0.025,
  test = "rmst difference",
  tau = NULL
)
}
\arguments{
\item{enrollRates}{enrollment rates}

\item{failRates}{failure and dropout rates}

\item{analysisTimes}{Minimum time of analysis}

\item{ratio}{Experimental:Control randomization ratio}

\item{alpha}{One-sided Type I error (strictly between 0 and 1)}

\item{test}{A string specifies the type of statistical test.
Default is \code{"survival difference"} (a Kaplan-Meier based test).
One can also set it as \code{"rmst difference"} (another Kaplan-Meier based test)}

\item{tau}{desired milestone for \code{test = "survival difference"} or \code{test = "rmst difference"}}
}
\description{
Power calculation based on RMST method
}
\examples{
# set enrollment rates
enrollRates <- tibble::tibble(Stratum = "All", duration = 12, rate = 500/12)

# set failure rates
failRates <- tibble::tibble(
  Stratum = "All",
  duration = c(4, 100),
  failRate = log(2) / 15,  # median survival 15 month
  hr = c(1, .6),
  dropoutRate = 0.001)

fixed_design_power_rmst(enrollRates, failRates, analysisTimes = 36)
fixed_design_power_rmst(enrollRates, failRates, analysisTimes = 36, tau = 18)
}
