% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_design_rmst.R
\name{fixed_design_size_rmst}
\alias{fixed_design_size_rmst}
\title{Sample Size Calculation based on RMST method}
\usage{
fixed_design_size_rmst(
  enrollRates,
  failRates,
  analysisTimes,
  ratio = 1,
  alpha = 0.025,
  beta = 0.1,
  tau = NULL
)
}
\arguments{
\item{enrollRates}{enrollment rates}

\item{failRates}{failure and dropout rates}

\item{analysisTimes}{Minimum time of analysis}

\item{ratio}{Experimental:Control randomization ratio}

\item{alpha}{One-sided Type I error (strictly between 0 and 1)}

\item{beta}{Power (\code{NULL} to compute power or strictly between 0 and \code{1 - alpha} otherwise)}

\item{tau}{desired milestone}
}
\value{
a list with \code{enrollRates}, \code{failRates}, \code{bounds}, \code{analysis} and \code{design}
}
\description{
Sample Size Calculation based on RMST method
}
\examples{
# set enrollment rates
enrollRates <- tibble::tibble(Stratum = "All", duration = 12, rate = 500/12)

# set failure rates
failRates <- tibble::tibble(
  Stratum = "All",
  duration = c(4, 100),
  failRate = log(2) / 15,  # median survival 15 month
  hr = c(1, .6),
  dropoutRate = 0.001)

fixed_design_size_rmst(enrollRates, failRates, analysisTimes = 36)
fixed_design_size_rmst(enrollRates, failRates, analysisTimes = 36, beta = 1 - 0.887)
fixed_design_size_rmst(enrollRates, failRates, analysisTimes = 36, tau = 18)
}
