% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gridptsRcpp}
\alias{gridptsRcpp}
\title{Grid points for group sequential design numerical integration in C++}
\usage{
gridptsRcpp(r, mu, a, b)
}
\arguments{
\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{mu}{Mean of normal distribution (scalar) under consideration}

\item{a}{lower limit of integration (scalar)}

\item{b}{upper limit of integration (scalar \code{> a})}
}
\value{
A \code{list} with grid points in \code{z} and numerical integration weights in \code{w}
}
\description{
Grid points for group sequential design numerical integration in C++
}
