% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_bound.R
\name{gs_bound}
\alias{gs_bound}
\title{Lower and Upper Bound of Group Sequential Design}
\usage{
gs_bound(
  alpha,
  beta,
  theta,
  corr,
  analysis = 1:length(alpha),
  theta0 = rep(0, length(analysis)),
  binding_lower_bound = FALSE,
  algorithm = GenzBretz(maxpts = 1e+05, abseps = 1e-05),
  alpha_bound = FALSE,
  beta_bound = FALSE,
  ...
)
}
\arguments{
\item{alpha}{a numeric vector of cumulative allocated alpha in each interim analysis}

\item{beta}{a numeric vector of cumulative allocated beta in each interim analysis}

\item{theta}{a numeric vector of effect size under alternative.}

\item{corr}{a matrix of correlation matrix}

\item{analysis}{a numeric vector of interim analysis indicator. Default is 1:length(alpha).}

\item{theta0}{a numeric vector of effect size under null hypothesis. Default is 0.}

\item{binding_lower_bound}{a logical value to indicate binding lower bound.}

\item{algorithm}{ an object of class \code{\link[mvtnorm]{GenzBretz}},
                    \code{\link[mvtnorm]{Miwa}} or \code{\link[mvtnorm]{TVPACK}}
                    specifying both the algorithm to be used as well as
                    the associated hyper parameters.}

\item{alpha_bound}{logical value to indicate if alpha is Type I error or upper bound. Default is FALSE.}

\item{beta_bound}{logical value to indicate if beta is Type II error or lower bound. Default is FALSE.}

\item{...}{additional parameters transfer to \code{mvtnorm::pmvnorm}}
}
\description{
Lower and Upper Bound of Group Sequential Design
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Create a vector of allocated alpha in each interim analysis from the cumulative allocated alpha.
   \item Create a vector of allocated beta in each interim analysis from the cumulative allocated beta.
   \item Extract the number of analysis.
   \item Find the upper and lower bound by solving multivariate normal distribution using \code{pmvnorm_combo}
   \item
   \item Return a data frame of upper and lower boundaries of group sequential design.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
library(gsDesign)

x <- gsDesign::gsSurv( k = 3 , test.type = 4 , alpha = 0.025 ,
                       beta = 0.2 , astar = 0 , timing = c( 1 ) ,
                       sfu = sfLDOF , sfupar = c( 0 ) , sfl = sfLDOF ,
                       sflpar = c( 0 ) , lambdaC = c( 0.1 ) ,
                       hr = 0.6 , hr0 = 1 , eta = 0.01 ,
                       gamma = c( 10 ) ,
                       R = c( 12 ) , S = NULL ,
                       T = 36 , minfup = 24 , ratio = 1 )

cbind(x$lower$bound, x$upper$bound)

gsdmvn:::gs_bound(alpha = sfLDOF(0.025, 1:3/3)$spend,
         beta = sfLDOF(0.2, 1:3/3)$spend,
         analysis = 1:3,
         theta = x$theta[2] * sqrt(x$n.I),
         corr = outer(1:3, 1:3, function(x,y) pmin(x,y) / pmax(x,y)))

}
