% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_design_rd.R
\name{gs_design_rd}
\alias{gs_design_rd}
\title{Group sequential design using average hazard ratio under non-proportional hazards}
\usage{
gs_design_rd(
  p_c = 0.15,
  p_e = 0.13,
  N = c(2, 4, 6),
  theta0 = 0,
  delta0 = 0,
  alpha = 0.025,
  beta = 0.1,
  ratio = 1,
  upper = gs_b,
  lower = gs_b,
  upar = list(par = gsDesign(k = length(N), test.type = 1, sfu = sfLDOF, sfupar =
    NULL)$upper$bound),
  lpar = list(par = c(qnorm(0.1), rep(-Inf, length(N) - 1))),
  test_upper = TRUE,
  test_lower = TRUE,
  info_scale = c(0, 1, 2),
  binding = FALSE,
  r = 18,
  tol = 1e-06
)
}
\arguments{
\item{p_c}{rate at the control group}

\item{p_e}{rate at the experimental group}

\item{theta0}{the standardized treatment effect under H0}

\item{delta0}{treatment effect under super-superiority designs, the default is 0}

\item{alpha}{One-sided Type I error}

\item{beta}{Type II error}

\item{ratio}{Experimental:Control randomization ratio (not yet implemented)}

\item{upper}{Function to compute upper bound}

\item{lower}{Function to compute lower bound}

\item{upar}{Parameter passed to \code{upper()}}

\item{lpar}{Parameter passed to \code{lower()}}

\item{test_upper}{indicator of which analyses should include an upper (efficacy) bound; single value of TRUE (default) indicates all analyses;
otherwise, a logical vector of the same length as \code{info} should indicate which analyses will have an efficacy bound}

\item{test_lower}{indicator of which analyses should include an lower bound; single value of TRUE (default) indicates all analyses;
single value FALSE indicated no lower bound; otherwise, a logical vector of the same length as \code{info} should indicate which analyses will have a
lower bound}

\item{info_scale}{}

\item{binding}{indicator of whether futility bound is binding; default of FALSE is recommended}

\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{tol}{Tolerance parameter for boundary convergence (on Z-scale)}

\item{n}{sample size}
}
\value{
a \code{tibble} with columns Analysis, Bound, Z, Probability, theta, Time, AHR, Events
}
\description{
Group sequential design using average hazard ratio under non-proportional hazards
}
\details{
Need to be added
}
\examples{
gs_design_rd()
}
