% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_info_rd.r
\name{gs_info_rd}
\alias{gs_info_rd}
\title{Information and effect size under risk difference}
\usage{
gs_info_rd(
  p_c = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3), Analysis =
    rep(1:3, 3), Rate = c(0.15, 0.2, 0.25, 0.13, 0.15, 0.18, 0.17, 0.22, 0.28)),
  p_e = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3), Analysis =
    rep(1:3, 3), Rate = c(0.12, 0.18, 0.21, 0.1, 0.11, 0.12, 0.11, 0.19, 0.26)),
  N = tibble::tibble(Stratum = rep(c("S1", "S2", "S3"), each = 3), Analysis = rep(1:3,
    3), N = c(50, 100, 200, 40, 80, 160, 60, 120, 240)),
  theta0 = 0,
  delta0 = 0,
  ratio = 1,
  weight = c("ss", "invar")
)
}
\arguments{
\item{p_c}{rate at the control group}

\item{p_e}{rate at the experimental group}

\item{theta0}{the standardized treatment effect under H0}

\item{delta0}{the treatment effect under H0}

\item{ratio}{Experimental:Control randomization ratio}

\item{n}{sample size}
}
\description{
Information and effect size under risk difference
}
\examples{
# --------------------- #
#      example 1        #
# --------------------- # 
# the default output is un-stratified
gs_info_rd()

# --------------------- #
#      example 2        #
# --------------------- #
# un-stratified case with H0: theta = 0
gs_info_rd(
  p_c = tibble::tibble(Stratum = "All", 
                       Rate = c(.15, .18, .2), 
                       Analysis = 1:3),
  p_e = tibble::tibble(Stratum = "All",
                       Rate = c(.12, .13, .15),
                       Analysis = 1:3),
  N = tibble::tibble(Stratum = "All",  
                     N = c(100, 200, 300),
                     Analysis = 1:3),
 strata = NULL,
 theta0 = 0,
 delta0 = 0, 
 ratio = 1
 )
 
# --------------------- #
#      example 3        #
# --------------------- #
# un-stratified case with H0: theta != 0
gs_info_rd(
  p_c = tibble::tibble(Stratum = "All", 
                       Rate = c(.15, .18, .2), 
                       Analysis = 1:3),
  p_e = tibble::tibble(Stratum = "All",
                       Rate = c(.12, .13, .15),
                       Analysis = 1:3),
  N = tibble::tibble(Stratum = "All",  
                     N = c(100, 200, 300),
                     Analysis = 1:3),
 strata = NULL,
 theta0 = 0.05,
 delta0 = 0.005, 
 ratio = 1
 )
 
# --------------------- #
#      example 4        #
# --------------------- #
# stratified case 
gs_info_rd(
  p_c = c(.15, .2, .25), 
  p_e = c(.1, .15, .2)
  N = tibble::tibble(Stratum = rep(c("A", "B", "C"), 3),  
                     N = rep(c(50, 100, 200), each = 3),
                     Analysis = rep(1:3, each = 3)),
  theta0 = 0,
  delta0 = 0, 
  ratio = 1)
}
