% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_power_rd.R
\name{gs_power_rd}
\alias{gs_power_rd}
\title{Title}
\usage{
gs_power_rd(
  p_c = 0.15,
  p_e = 0.1,
  n = c(350, 700, 1400),
  delta = NULL,
  delta0 = 0,
  ratio = 1,
  upper = gs_b,
  upar = list(par = gsDesign(k = length(n), test.type = 1, sfu = sfLDOF, sfupar =
    NULL)$upper$bound),
  lower = gs_b,
  lpar = list(par = c(qnorm(0.1), rep(-Inf, length(n) - 1))),
  info_scale = c(0, 1, 2),
  binding = FALSE,
  test_upper = TRUE,
  test_lower = TRUE,
  r = 18,
  tol = 1e-06
)
}
\arguments{
\item{p_c}{}

\item{p_e}{}

\item{n}{sample size}

\item{delta}{}

\item{delta0}{treatment effect under super-superiority designs, the default is 0}

\item{ratio}{experimental:control randomization ratio}

\item{upper}{function to compute upper bound}

\item{upar}{parameter to pass to upper}

\item{lower}{function to compare lower bound}

\item{lpar}{parameter to pass to lower}

\item{info_scale}{}

\item{binding}{indicator of whether futility bound is binding; default of FALSE is recommended}

\item{test_upper}{indicator of which analyses should include an upper (efficacy) bound;
single value of TRUE (default)  indicates all analyses; otherwise,
a logical vector of the same length as \code{info} should indicate which analyses will have an efficacy bound}

\item{test_lower}{indicator of which analyses should include a lower bound;
single value of TRUE (default) indicates all analyses;
single value FALSE indicated no lower bound; otherwise,
a logical vector of the same length as \code{info} should indicate which analyses will have a lower bound}

\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{tol}{Tolerance parameter for boundary convergence (on Z-scale)}
}
\value{

}
\description{
Title
}
\examples{
gs_power_rd()
}
