% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_prob.r
\name{gs_prob}
\alias{gs_prob}
\title{Group sequential boundary crossing probabilities}
\usage{
gs_prob(theta, upper = gs_b, lower = gs_b, upar, lpar, info, r = 18)
}
\arguments{
\item{theta}{natural parameter for group sequentia design representing expected drift at time of each analysis}

\item{upper}{function to compute upper bound}

\item{lower}{function to compare lower bound}

\item{upar}{parameter to pass to upper}

\item{lpar}{parameter to pass to lower}

\item{info}{statistical information at each analysis}

\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}
}
\value{
A `tibble` with a row for each finite bound and analysis containing the following variables:
Analysis analysis number
Bound Upper (efficacy) or Lower (futility)
Z Z-value at bound
Probability probability that this is the first bound crossed under the given input
theta approximate natural parameter value required to cross the bound
}
\description{
Group sequential boundary crossing probabilities
}
\details{
Approximation for \code{theta} is based on Wald test and assumes the observed information is equal to the expected.
}
\examples{
# Asymmetric 2-sided design
gs_prob(theta = 0, upar = rep(2.2, 3), lpar = rep(0, 3), upper=gs_b, lower=gs_b,  info = 1:3)
# One-sided design
x <- gs_prob(theta = 0, upar = rep(2.2, 3), lpar = rep(-Inf, 3), upper=gs_b, lower=gs_b,  info = 1:3)
# Without filtering, this shows unneeded lower bound
x
# Filter to just show bounds intended for use
x \%>\% filter(abs(Z) < Inf)
}
