% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h1Rcpp}
\alias{h1Rcpp}
\title{Initialize numerical integration for group sequential design in C++}
\usage{
h1Rcpp(r, theta, I, a, b)
}
\arguments{
\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{theta}{Drift parameter for first analysis}

\item{I}{Information at first analysis}

\item{a}{lower limit of integration (scalar)}

\item{b}{upper limit of integration (scalar \code{> a})}
}
\value{
A \code{tibble} with grid points in \code{z}, numerical integration weights in \code{w},
and a normal density with mean \code{mu = theta * sqrt{I}} and variance 1 times the weight in \code{w}.
}
\description{
Compute grid points for first interim analysis in a group sequential design
}
