% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hupdate_.r
\name{hupdate_}
\alias{hupdate_}
\title{Update numerical integration for group sequential design}
\usage{
hupdate_(
  r = 18,
  theta = 0,
  I = 2,
  a = -Inf,
  b = Inf,
  thetam1 = 0,
  Im1 = 1,
  gm1 = h1_()
)
}
\arguments{
\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{theta}{Drift parameter for current analysis}

\item{I}{Information at current analysis}

\item{a}{lower limit of integration (scalar)}

\item{b}{upper limit of integration (scalar \code{> a})}

\item{thetam1}{Drift parameter for previous analysis}

\item{Im1}{Information at previous analysis}

\item{gm1}{numerical integration grid from \code{h1()} or previous run of \code{hupdate()}}
}
\value{
A \code{tibble} with grid points in \code{z}, numerical integration weights in \code{w},
and a normal density with mean \code{mu = theta * sqrt{I}} and variance 1 times the weight in \code{w}.
}
\description{
Update grid points for numerical integration from one analysis to the next
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Compute the square root of the change in information.
   \item Compute the grid points for group sequential design numerical integration.
   \item Update the integration.
   \item Return a tibble of z, w, and h.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
library(dplyr)
# 2nd analysis with no interim bound and drift 0 should have mean 0, variance 1
hupdate_() \%>\% summarise(mu = sum(z * h), var = sum((z - mu)^2 * h))
}
