% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hupdateRcpp}
\alias{hupdateRcpp}
\title{Update numerical integration for group sequential design in C++}
\usage{
hupdateRcpp(r, theta, I, a, b, thetam1, Im1, gm1)
}
\arguments{
\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{theta}{Drift parameter for current analysis}

\item{I}{Information at current analysis}

\item{a}{lower limit of integration (scalar)}

\item{b}{upper limit of integration (scalar \code{> a})}

\item{thetam1}{Drift parameter for previous analysis}

\item{Im1}{Information at previous analysis}

\item{gm1}{numerical integration grid from \code{h1()} or previous run of \code{hupdate()}}
}
\value{
A \code{tibble} with grid points in \code{z}, numerical integration weights in \code{w},
and a normal density with mean \code{mu = theta * sqrt{I}} and variance 1 times the weight in \code{w}.
}
\description{
Update grid points for numerical integration from one analysis to the next
}
