% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fixed_design}
\alias{summary.fixed_design}
\title{summary for \code{fixed_design()} object}
\usage{
\method{summary}{fixed_design}(x, ...)
}
\arguments{
\item{x}{a fixed design object returned by \code{fixed_design()}}

\item{...}{additional arguments}
}
\description{
summary function to fixed_design class
}
\examples{
fixed_design(x = "AHR", 
             alpha = alpha,  
             enrollRates = enrollRates, 
             failRates = failRates, 
             studyDuration = studyDuration, 
             ratio = ratio
             ) \%>\% 
             summary()
}
