% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_design_rmst.R
\name{fixed_design_power_rmst}
\alias{fixed_design_power_rmst}
\title{Power calculation based on RMST method}
\usage{
fixed_design_power_rmst(
  enrollRates,
  failRates,
  analysisTimes,
  ratio = 1,
  alpha = 0.025
)
}
\description{
Power calculation based on RMST method
}
\examples{
# set enrollment rates
enrollRates <- tibble::tibble(Stratum = "All", duration = 12, rate = 500/12)

# set failure rates
failRates <- tibble::tibble(
  Stratum = "All",
  duration = c(4, 100),
  failRate = log(2) / 15,  # median survival 15 month
  hr = c(1, .6),
  dropoutRate = 0.001)

fixed_design_power_rmst(enrollRates, failRates, analysisTimes = 36)
fixed_design_power_rmst(enrollRates, failRates, analysisTimes = 36)
fixed_design_power_rmst(enrollRates, failRates, analysisTimes = 36, beta = 1 - 0.887)
}
