% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_info_rd.r
\name{gs_info_rd}
\alias{gs_info_rd}
\title{Information and effect size under risk difference}
\usage{
gs_info_rd(
  p_c = 0.15,
  p_e = 0.13,
  n = c(50, 80, 100),
  theta0 = 0,
  delta0 = 0,
  ratio = 1
)
}
\arguments{
\item{p_c}{rate at the control group}

\item{p_e}{rate at the experimental group}

\item{n}{sample size}

\item{theta0}{the standardized treatment effect under H0}

\item{delta0}{the treatment effect under H0}

\item{ratio}{Experimental:Control randomization ratio}
}
\description{
Information and effect size under risk difference
}
\examples{
gs_info_rd()

}
