% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_power_nph.r
\name{gs_power_nph}
\alias{gs_power_nph}
\title{Group sequential design power under non-proportional hazards}
\usage{
gs_power_nph(
  enrollRates = tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6,
    9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  ratio = 1,
  events = c(30, 40, 50),
  analysisTimes = NULL,
  maxEvents = 45,
  upper = gs_b,
  upar = gsDesign(k = length(events), test.type = 1, n.I = events, maxn.IPlan =
    maxEvents, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lower = gs_b,
  lpar = c(qnorm(0.1), rep(-Inf, 2)),
  r = 18
)
}
\arguments{
\item{enrollRates}{enrollment rates}

\item{failRates}{failure and dropout rates}

\item{ratio}{Experimental:Control randomization ratio (not yet implemented)}

\item{events}{Targeted events at each analysis}

\item{analysisTimes}{Not yet implemented}

\item{maxEvents}{Final planned events}

\item{upper}{Function to compute upper bound}

\item{upar}{Parameter passed to \code{upper()}}

\item{lower}{Function to compute lower bound}

\item{lpar}{Parameter passed to \code{lower()}}

\item{r}{Control for grid size; normally leave at default of \code{r=18}}
}
\value{
a \code{tibble} with columns Analysis, Bound, Z, Probability, theta, Time, avehr, Events
}
\description{
Group sequential design power under non-proportional hazards
}
\details{
Need to be added
}
\examples{
library(gsDesign)
library(gsDesign2)
gs_power_nph() \%>\% filter(abs(Z) < Inf)
}
