% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridptsRcpp.R
\name{h1Rcpp}
\alias{h1Rcpp}
\title{Initialize numerical integration for group sequential design}
\usage{
h1Rcpp(r = 18, theta = 0, I = 1, a = -Inf, b = Inf)
}
\arguments{
\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{theta}{Drift parameter for first analysis}

\item{I}{Information at first analysis}

\item{a}{lower limit of integration (scalar)}

\item{b}{upper limit of integration (scalar \code{> a})}
}
\value{
A \code{list} with grid points in \code{z}, numerical integration weights in \code{w},
and a normal density with mean \code{mu = theta * sqrt{I}} and variance 1 times the weight in \code{w}.
}
\description{
Compute grid points for first interim analysis in a group sequential design
}
\details{
Mean for standard normal distribution under consideration is \code{mu = theta * sqrt(I)}
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Compute drift at analysis 1.
   \item Compute deviation from drift.
   \item Compute standard normal density, multiply by grid weight.
   \item Return a tibble of z, w, and h.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

